WindowingGUI - Toolbar Menu V0.3
Update to the main app so that we return Logs for any posible borked behaviors. Important as we update this app
